:recycle: Move TestFixtureProvider one level up
